Add validation to single photo download
