--- backend/src/controlmat.Application/Common/Queries/Photo/DownloadPhotoQuery.cs
@@ -1,59 +1,0 @@
-using MediatR;
-using Microsoft.Extensions.Logging;
-using Controlmat.Application.Common.Dto;
-using Controlmat.Domain.Interfaces;
-using System.IO;
-
-namespace Controlmat.Application.Common.Queries.Photo;
-
-public static class DownloadPhotoQuery
-{
-    public record Request(int PhotoId) : IRequest<PhotoDownloadDto?>;
-
-    public class Handler : IRequestHandler<Request, PhotoDownloadDto?>
-    {
-        private readonly IPhotoRepository _repository;
-        private readonly ILogger<Handler> _logger;
-
-        public Handler(IPhotoRepository repository, ILogger<Handler> logger)
-        {
-            _repository = repository;
-            _logger = logger;
-        }
-
-        public async Task<PhotoDownloadDto?> Handle(Request request, CancellationToken ct)
-        {
-            _logger.LogInformation("🌀 DownloadPhotoQuery - STARTED. PhotoId: {PhotoId}", request.PhotoId);
-
-            var photo = await _repository.GetByIdAsync(request.PhotoId);
-            if (photo == null)
-            {
-                _logger.LogWarning("⚠️ Photo not found: {PhotoId}", request.PhotoId);
-                return null;
-            }
-
-            if (!File.Exists(photo.FilePath))
-            {
-                _logger.LogError("❌ Photo file not found on disk: {FilePath}", photo.FilePath);
-                return null;
-            }
-
-            var fileBytes = await File.ReadAllBytesAsync(photo.FilePath, ct);
-            var contentType = Path.GetExtension(photo.FileName).ToLower() switch
-            {
-                ".jpg" or ".jpeg" => "image/jpeg",
-                ".png" => "image/png",
-                _ => "application/octet-stream"
-            };
-
-            _logger.LogInformation("✅ DownloadPhotoQuery - COMPLETED. File size: {Size} bytes", fileBytes.Length);
-
-            return new PhotoDownloadDto
-            {
-                FileBytes = fileBytes,
-                ContentType = contentType,
-                FileName = photo.FileName
-            };
-        }
-    }
-}
