+++ backend/src/controlmat.Domain/Interfaces/IPhotoRepository.cs
@@ -7,6 +7,7 @@ namespace Controlmat.Domain.Interfaces;
 
 public interface IPhotoRepository
 {
+    Task<bool> ExistsAsync(int photoId);
     Task<Photo?> GetByIdAsync(int photoId);
     Task<List<Photo>> GetByWashingIdAsync(long washingId);
     Task AddAsync(Photo photo);
